Add gesture based lockscreen security. [1/2]
This adds the ability to us a hand drawn gesture as
a means to unlock the device.

In the case that the gesture is forgotton or not
being recognized correctly, the fallback will be the
same as when using patterns.

PS2:
Return SecurityMode.Account as backup for gesture login.

PS3:
Poke the wakelock when a gesture is detected so the screen
will not timeout when an invalid gesture is entered.

PS4:
Clean things up a bit.

Change-Id: I4ca8d7edff0a8906220cf4561d8b90896520e0cd

Conflicts:
	core/java/com/android/internal/widget/LockPatternUtils.java
	core/java/com/android/internal/widget/LockSettingsService.java